backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
python-dotenv==1.0.0
google-generativeai==0.3.1
python-multipart==0.0.6
pydantic==2.5.1